Add files via upload
Mechanism of Action Prediction
What is Mechanim of Action?
The term mechanism of action means the biochemical interactions through which a drug generates its pharmacological effect.

Why to predict Mechanism of Action ?
If we know a disease affects some particular receptor or downstream set of cell activity, we can develop drugs faster if we can predict how cells and genes affect various receptor sites.

Problem Discription
For this competition,we will have dataset that combines gene expression and cell viability data in addition to the MoA annotations of more than 5,000 drugs. In this competition, each drug was tested under two dose (cp_dose) and three times (cp_time). So, six samples basically correspond to one drug. We are going to train a model that classifies drugs based on their biological activity. This problem is a multilabel classification, which means we have multiple targets (not multiple classes). In this project, we will first perform explanatory data analysis and then train a model using deep neural networks with Keras. We will do a bit model evaluation at the end.